Improve Railway status output
