+++ src/commands/output/mod.rs
@@ -1,0 +1,1 @@
+pub(in crate::commands) mod service_summary;
